Return investor Id in response

+++ Backend/InvestorCommitments/InvestorCommitments.Infrastructure/Repository/Models/InvestorDto.cs
@@ -6,5 +6,5 @@ public record InvestorDto
     public required string Name { get; set; }
     public required string InvestoryType { get; set; }
     public required string Country { get; set; }
-    public required int TotalCommitments { get; set; }
+    public required double  TotalCommitments { get; set; }
 }
